requirements.txt
grpcio
grpcio-tools
msgpack
protobuf
fastapi
uvicorn
pydantic
psycopg2-binary
PyJWT[crypto]
pytest